coudlnt upload more comments to discussion page
my goal was to use multithreading inorder to run the multiple ui windows, client server, and mysql database.

once the server is open, the client would open which would also run the login page, file and the myjdbc file.

from here i was going to have the information the user input get sent from loginpage ui, to the user class, where it would look at the database and compare if the user information matches. there we can log in.

the other pages would also have their own threads inorder to run independantly from the loginpage unlike my april project. in my april project, all my ui pages would be ran through the one file. this would be messy since it created a large file. it was also tricky on how to have the pages show up separately. i figured to have actionlisteners change the status of visability for each page. the biggest problem with the old setup was if i were to close the forget password, as an example, it would close out all the other windows.

you can tell the code is pretty different between may and april. I could have just implemented databases, lambdas, concurencies, and stream api directly to the old one but i found it way too messy with all the multidimensional arrays. 

I didnt get around to making a page where once the user logged in, the user would be able to interact with a list of users. sorting them by gpa, alphabetical order, and a few other criteria.

i also wanted to create a window where it would display the user's account with all their public information. i wanted to create a io stream to allow the user to download their profile to a txt file. 